[TNT-151] move: 트레이니 회원가입 화면 trainee:signup 모듈로 이동
