+++ core/designsystem/src/main/res/values/strings.xml
@@ -5,46 +5,12 @@
     <string name="placeholder_content_input">내용을 입력해주세요</string>
 
     <!--  Sign Up  -->
-    <string name="name">이름</string>
     <string name="next">다음</string>
     <string name="start">시작하기</string>
     <string name="trainer">트레이너</string>
     <string name="trainee">트레이니</string>
 
-    <string name="birthday_label">생년월일</string>
-    <string name="height_label">키</string>
-    <string name="weight_label">체중</string>
-    <string name="height_unit">cm</string>
-    <string name="weight_unit">kg</string>
-
-    <string name="signup_warning_text_length">자 이하로 입력해주세요</string>
-
     <string name="signup_select_role_title">안녕하세요!\n어떤 회원으로 이용하시겠어요?</string>
     <string name="signup_select_role_subtitle">트레이너와 트레이니 중 선택해주세요.</string>
-
-    <string name="signup_set_name_title">이름이 어떻게 되세요?</string>
-    <string name="signup_set_name_placeholder">이름을 입력해주세요</string>
-
-    <string name="signup_set_basic_info_title">회원님의 기본 정보를\n입력해주세요</string>
-    <string name="signup_set_basic_info_subtitle">PT에 참고할 기본 정보예요!</string>
-    <string name="signup_birthday_placeholder">2001/01/01</string>
-
-    <string name="signup_pt_purpose_title">PT를 받는 목적에 대해\n알려주세요!</string>
-    <string name="signup_pt_purpose_subtitle">다중 선택이 가능해요.</string>
-    <string name="signup_pt_purpose_loss_weight">체중 감량</string>
-    <string name="signup_pt_purpose_strength">근력 향상</string>
-    <string name="signup_pt_purpose_health_care">건강 관리</string>
-    <string name="signup_pt_purpose_flexibility">유연성 향상</string>
-    <string name="signup_pt_purpose_body_profile">바디프로필</string>
-    <string name="signup_pt_purpose_posture_correction">자세 교정</string>
-
-    <string name="signup_pt_precaution_title">트레이너가 꼭 알아야 할\n주의사항이 있나요?</string>
-    <string name="signup_pt_precaution_subtitle">트레이너에게 알려드릴게요.</string>
-
-    <string name="signup_complete_trainer_title">만나서 반가워요\n%s 트레이너님!</string>
-    <string name="signup_complete_trainer_subtitle">트레이니와 함께\n케미를 터뜨려보세요! 🧨</string>
-
-    <string name="signup_complete_trainee_title">만나서 반가워요\n%s 트레이니님!</string>
-    <string name="signup_complete_trainee_subtitle">트레이너와 함께\n케미를 터뜨려보세요! 🧨</string>
 
 </resources>
