+++ feature/trainee/signup/src/main/res/values/strings.xml
@@ -1,0 +1,37 @@
+<?xml version="1.0" encoding="utf-8"?>
+<resources>
+    <string name="next">다음</string>
+    <string name="start">시작하기</string>
+
+
+    <string name="name">이름</string>
+    <string name="birthday_label">생년월일</string>
+    <string name="height_label">키</string>
+    <string name="weight_label">체중</string>
+    <string name="height_unit">cm</string>
+    <string name="weight_unit">kg</string>
+
+    <string name="what_is_your_name">이름이 어떻게 되세요?</string>
+    <string name="enter_your_name">이름을 입력해주세요</string>
+    <string name="text_length_warning">자 이하로 입력해주세요</string>
+
+    <string name="enter_your_basic_info">회원님의 기본 정보를\n입력해주세요</string>
+    <string name="basic_info_for_pt">PT에 참고할 기본 정보예요!</string>
+    <string name="birthday_placeholder">2001/01/01</string>
+
+    <string name="purpose_of_pt">PT를 받는 목적에 대해\n알려주세요!</string>
+    <string name="multiple_choices_available">다중 선택이 가능해요.</string>
+    <string name="loss_weight">체중 감량</string>
+    <string name="strength_improvement">근력 향상</string>
+    <string name="health_care">건강 관리</string>
+    <string name="flexibility">유연성 향상</string>
+    <string name="body_profile">바디프로필</string>
+    <string name="posture_correction">자세 교정</string>
+
+    <string name="caution_that_trainer_must_know">트레이너가 꼭 알아야 할\n주의사항이 있나요?</string>
+    <string name="let_the_trainer_know">트레이너에게 알려드릴게요.</string>
+
+    <string name="nice_to_meet_you_trainee">만나서 반가워요\n%s 트레이니님!</string>
+    <string name="chemistry_boom_with_trainer">트레이너와 함께\n케미를 터뜨려보세요! 🧨</string>
+
+</resources>
